+++ src/main/java/com/bho/catchtrippingbackend/user/dto/UserRegisterResponseDto.java
@@ -1,5 +1,5 @@
 package com.bho.catchtrippingbackend.user.dto;
 
 public record UserRegisterResponseDto(
-    int userId
+    Long userId
 ) {}
